Fixed warnings + relabel of nodes

+++ src/Bpp/Phyl/Likelihood/DataFlow/LikelihoodCalculationOnABranch.h
@@ -205,7 +205,7 @@ public:
     {
       auto pos = Eigen::Index(currentPosition);
       if (pos>=rootPatternLinks_->targetValue().rows())
-        throw BadSizeException("Forbidden access in getRootArrayPosition.",pos,rootPatternLinks_->targetValue().rows());
+        throw BadSizeException("Forbidden access in getRootArrayPosition.",static_cast<size_t>(pos),static_cast<size_t>(rootPatternLinks_->targetValue().rows()));
       else
         return rootPatternLinks_->targetValue()(pos);
     }
